Рефакторинг
Раскидал все по отдельным заголовочным файлам
