+++ Kursach/Constants.h
@@ -1,0 +1,5 @@
+#pragma once
+
+const int HEIGHT = 1024, WIDTH = 1024;
+const int HEIGHT_WINDOW = HEIGHT, WIDTH_WINDOW = WIDTH + 300;
+const float SEA_LEVEL = 0.6f, SAND_LEVEL = 0.62f, GRASS_LEVEL = 0.97f;
